scripts: bareboximd: remove usage of loff_t
loff_t is Linux-specific and unneeded at this place, as sizes are usually
denoted with a size_t. As the signedness of max_size isn't relied on
anywhere, replace it with a size_t.

+++ scripts/bareboximd.c
@@ -58,7 +58,7 @@ int imd_command_setenv(const char *variable_name, const char *value)
 	return -EINVAL;
 }
 
-static int read_file_2(const char *filename, size_t *size, void **outbuf, loff_t max_size)
+static int read_file_2(const char *filename, size_t *size, void **outbuf, size_t max_size)
 {
 	off_t fsize;
 	ssize_t rsize;
